+++ Sources/KeyboardKit/UIKit/Buttons/KeyboardButtonRow.swift
@@ -16,7 +16,7 @@ open class KeyboardButtonRow: UIView, KeyboardStackViewComponent {
     
     public convenience init(
         actions: KeyboardActionRow,
-        height: CGFloat = standardHeight(),
+        height: CGFloat = .standardKeyboardRowHeight(),
         alignment: UIStackView.Alignment = .fill,
         distribution: UIStackView.Distribution = .fillEqually,
         buttonCreator: KeyboardButtonCreator) {